+++ config/webpack/plugin/SetupPlugin.js
@@ -64,26 +64,11 @@ class SetupPlugin {
             // Use MDX to import any md files imported from an mdx file.
             issuer: [
               /\.mdx?$/,
-            ],
-            use: [
-              babelLoader,
-              mdxLoader,
-            ],
-          },
-          {
-            // Use the marked loader to load any md files included by the content config file.
-            // LOAD USING MDX ON NEXT MAJOR VERSION.
-            issuer: [
               /dev-site-config.*[/\\]contentConfig\.js$/,
             ],
             use: [
               babelLoader,
-              {
-                loader: 'devSiteMarked',
-                options: {
-                  baseUrl: this.publicPath,
-                },
-              },
+              mdxLoader,
             ],
           },
           {